refactor(scan) refactor PlanContext to allow for different manifest selections
